major steps toward understanding props and props data
