+++ src/components/Dice_Game/dicegame.css
@@ -3,7 +3,6 @@
     width: 200px;
     position: relative;
     transform-style: preserve-3d;
-    /* transform: translateZ(-100px) rotateY(-45deg) rotateX(-45deg); */
     transition: transform .8s;
     background-color: rgb(198, 198, 198);
 }
@@ -13,7 +12,7 @@
     width: 200px;
     height: 200px;
     perspective: 400px;
-    position: absolute;
+    position: relative;
     left: 50%;
     top: 30%;
     transform: translate(-50%, -50%);
@@ -115,29 +114,9 @@
     z-index: 6;
 }
 
-/*Customization panel*/
-.custom {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: calc(20% + 250px);
-}
-
-.custom>* {
-    margin: 10px;
-}
-
-.custom>button {
-    font-size: 20px;
-    font-family: "Verdana";
-    width: 200px;
-    height: 50px;
-    color: gold;
-    background-color: black;
-    border: none;
-    border-radius: 10px;
-}
-
-.picker {
-    border: none;
+.diceBtn {
+    background-color: rgb(24, 137, 218);
+    padding: 1em;
+    margin: 1em;
+    font-size: x-large;
 }
