+++ src/components/Dice_Game/sounds/resources.txt
@@ -1,0 +1,6 @@
+I am going to paste links to resources such as images and sounds in this file for now
+
+Sounds -
+https://freesound.org/s/318968/
+
+Images -
